requirement.txt
PyTorch==1.0.2
Numpy==1.15.4
Pillow==5.4.1
h5py==2.8.0
tqdm==4.30.0
